move various things out of lib.rs to more specific modules
